:bookmark_tabs: :: infrastructure 구조 수정
