+++ feed-infrastructure/src/main/java/com/xquare/v1servicefeed/attachment/domain/AttachmentEntityId.java
@@ -1,0 +1,29 @@
+package com.xquare.v1servicefeed.attachment.domain;
+
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.experimental.SuperBuilder;
+import org.hibernate.annotations.GenericGenerator;
+
+import javax.persistence.Column;
+import javax.persistence.Embeddable;
+import javax.persistence.GeneratedValue;
+import java.io.Serializable;
+import java.util.UUID;
+
+@Getter
+@SuperBuilder
+@NoArgsConstructor
+@EqualsAndHashCode
+@Embeddable
+public class AttachmentEntityId implements Serializable {
+
+    @GeneratedValue(generator = "uuid2")
+    @GenericGenerator(name = "uuid2", strategy = "uuid2")
+    @Column(columnDefinition = "BINARY(16)")
+    private UUID id;
+
+    @Column(nullable = false)
+    private UUID feedId;
+}
