+++ feed-infrastructure/src/main/java/com/xquare/v1servicefeed/attachment/domain/mapper/AttachmentMapper.java
@@ -1,0 +1,4 @@
+package com.xquare.v1servicefeed.attachment.domain.mapper;
+
+public class AttachmentMapper {
+}
