+++ feed-infrastructure/src/main/java/com/xquare/v1servicefeed/attachment/domain/repository/AttachmentRepositoryAdapter.java
@@ -1,0 +1,11 @@
+package com.xquare.v1servicefeed.attachment.domain.repository;
+
+import com.xquare.v1servicefeed.configuration.annotation.Adapter;
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+@Adapter
+public class AttachmentRepositoryAdapter {
+
+    private final AttachmentRepository attachmentRepository;
+}
