+++ feed-infrastructure/src/main/java/com/xquare/v1servicefeed/comment/domain/mapper/AttachmentMapper.java
@@ -1,0 +1,4 @@
+package com.xquare.v1servicefeed.comment.domain.mapper;
+
+public class AttachmentMapper {
+}
