+++ feed-infrastructure/src/main/java/com/xquare/v1servicefeed/comment/web/WebCommentAdapter.java
@@ -1,0 +1,9 @@
+package com.xquare.v1servicefeed.comment.web;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RestController;
+
+@RequiredArgsConstructor
+@RestController
+public class WebCommentAdapter {
+}
