--- feed-infrastructure/src/main/java/com/xquare/v1servicefeed/configuration/entity/BaseUUIDEntity.java
@@ -1,25 +1,0 @@
-package com.xquare.v1servicefeed.configuration.entity;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.experimental.SuperBuilder;
-import org.hibernate.annotations.GenericGenerator;
-
-import javax.persistence.Column;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.MappedSuperclass;
-import java.util.UUID;
-
-@Getter
-@SuperBuilder
-@NoArgsConstructor
-@MappedSuperclass
-public abstract class BaseUUIDEntity {
-
-    @Id
-    @GeneratedValue(generator = "uuid2")
-    @GenericGenerator(name = "uuid2", strategy = "uuid2")
-    @Column(columnDefinition = "BINARY(16)")
-    private UUID id;
-}
