--- feed-infrastructure/src/main/java/com/xquare/v1servicefeed/domain/comment/domain/CommentEntity.java
@@ -1,36 +1,0 @@
-package com.xquare.v1servicefeed.domain.comment.domain;
-
-import com.xquare.v1servicefeed.configuration.entity.BaseUUIDEntity;
-import com.xquare.v1servicefeed.domain.feed.domain.FeedEntity;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.experimental.SuperBuilder;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.Table;
-import java.util.UUID;
-
-@Getter
-@SuperBuilder
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "tbl_comment")
-@Entity
-public class CommentEntity extends BaseUUIDEntity {
-
-    @Column(nullable = false)
-    private String content;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "feed_id", nullable = false)
-    private FeedEntity feed;
-
-    @Column(columnDefinition = "BINARY(16)", nullable = false)
-    private UUID userId;
-}
