--- feed-infrastructure/src/main/java/com/xquare/v1servicefeed/domain/comment/domain/mapper/AttachmentMapper.java
@@ -1,4 +1,0 @@
-package com.xquare.v1servicefeed.domain.comment.domain.mapper;
-
-public class AttachmentMapper {
-}
