--- feed-infrastructure/src/main/java/com/xquare/v1servicefeed/domain/feed/domain/mapper/FeedMapper.java
@@ -1,7 +1,0 @@
-package com.xquare.v1servicefeed.domain.feed.domain.mapper;
-
-import org.springframework.stereotype.Component;
-
-@Component
-public class FeedMapper {
-}
