--- feed-infrastructure/src/main/java/com/xquare/v1servicefeed/domain/feed/web/WebFeedAdapter.java
@@ -1,9 +1,0 @@
-package com.xquare.v1servicefeed.domain.feed.web;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.RestController;
-
-@RequiredArgsConstructor
-@RestController
-public class WebFeedAdapter {
-}
