--- feed-infrastructure/src/main/java/com/xquare/v1servicefeed/domain/feedlike/domain/mapper/FeedLikeMapper.java
@@ -1,7 +1,0 @@
-package com.xquare.v1servicefeed.domain.feedlike.domain.mapper;
-
-import org.springframework.stereotype.Component;
-
-@Component
-public class FeedLikeMapper {
-}
