+++ feed-infrastructure/src/main/java/com/xquare/v1servicefeed/feed/domain/FeedEntity.java
@@ -1,0 +1,41 @@
+package com.xquare.v1servicefeed.feed.domain;
+
+import com.xquare.v1servicefeed.entity.BaseUUIDEntity;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.experimental.SuperBuilder;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.Table;
+import java.time.LocalDateTime;
+import java.util.UUID;
+
+@Getter
+@SuperBuilder
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Table(name = "tbl_feed")
+@Entity
+public class FeedEntity extends BaseUUIDEntity {
+
+    @Column(columnDefinition = "BINARY(16)", nullable = false)
+    private UUID userId;
+
+    @Column(nullable = false)
+    private String title;
+
+    @Column(columnDefinition = "TEXT", nullable = false)
+    private String content;
+
+    @Column(nullable = false)
+    private LocalDateTime createdAt;
+
+    @Column(nullable = false)
+    private Integer likeCount;
+
+    @Column(columnDefinition = "CHAR(6)", nullable = false)
+    private String category;
+}
