+++ feed-infrastructure/src/main/java/com/xquare/v1servicefeed/feed/domain/mapper/FeedMapper.java
@@ -1,0 +1,7 @@
+package com.xquare.v1servicefeed.feed.domain.mapper;
+
+import org.springframework.stereotype.Component;
+
+@Component
+public class FeedMapper {
+}
