+++ feed-infrastructure/src/main/java/com/xquare/v1servicefeed/feed/domain/repository/FeedRepositoryAdapter.java
@@ -1,0 +1,11 @@
+package com.xquare.v1servicefeed.feed.domain.repository;
+
+import com.xquare.v1servicefeed.configuration.annotation.Adapter;
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+@Adapter
+public class FeedRepositoryAdapter {
+
+    private final FeedRepository feedRepository;
+}
