+++ feed-infrastructure/src/main/java/com/xquare/v1servicefeed/feed/web/WebFeedAdapter.java
@@ -1,0 +1,9 @@
+package com.xquare.v1servicefeed.feed.web;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RestController;
+
+@RequiredArgsConstructor
+@RestController
+public class WebFeedAdapter {
+}
