+++ feed-infrastructure/src/main/java/com/xquare/v1servicefeed/feedlike/domain/mapper/FeedLikeMapper.java
@@ -1,0 +1,7 @@
+package com.xquare.v1servicefeed.feedlike.domain.mapper;
+
+import org.springframework.stereotype.Component;
+
+@Component
+public class FeedLikeMapper {
+}
