+++ feed-infrastructure/src/main/java/com/xquare/v1servicefeed/feedlike/domain/repository/FeedLikeRepositoryAdapter.java
@@ -1,0 +1,11 @@
+package com.xquare.v1servicefeed.feedlike.domain.repository;
+
+import com.xquare.v1servicefeed.configuration.annotation.Adapter;
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+@Adapter
+public class FeedLikeRepositoryAdapter {
+
+    private final FeedLikeRepository feedLikeRepository;
+}
