+++ feed-infrastructure/src/main/java/com/xquare/v1servicefeed/feedlike/web/WebFeedLikeAdapter.java
@@ -1,0 +1,9 @@
+package com.xquare.v1servicefeed.feedlike.web;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RestController;
+
+@RequiredArgsConstructor
+@RestController
+public class WebFeedLikeAdapter {
+}
